docs(button): add back-to-top buttons
Added a button for almost all sections, that navigates back to the table of contents. It seems that it would be helpful if a reader is interested in more than one section but they are separated. So, the reader can go back to the table of contents and click another section.